requirements.txt
appdirs==1.4.0
blinker==1.3
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
future==0.16.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
oauthlib==2.0.1
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.6.1
pyparsing==2.1.10
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
SQLAlchemy==1.0.3
TwitterAPI==2.4.3
Werkzeug==0.10.4
